move typescript resolvers to separate template added support for setting the context type in resolvers template
