+++ packages/templates/typescript-resolvers/.npmignore
@@ -1,0 +1,4 @@
+src
+node_modules
+tests
+!dist
