+++ packages/templates/typescript-resolvers/package.json
@@ -1,0 +1,44 @@
+{
+  "name": "graphql-codegen-typescript-resolvers-template",
+  "version": "0.12.6",
+  "description": "",
+  "license": "MIT",
+  "scripts": {
+    "prepublishOnly": "yarn build",
+    "build": "codegen-templates-scripts build",
+    "pretest": "yarn build",
+    "test": "codegen-templates-scripts test"
+  },
+  "devDependencies": {
+    "graphql-codegen-typescript-template": "0.12.6",
+    "codegen-templates-scripts": "0.12.6",
+    "graphql-codegen-core": "0.12.6",
+    "graphql-codegen-compiler": "0.12.6",
+    "graphql": "14.0.2",
+    "@types/graphql": "14.0.3"
+  },
+  "main": "./dist/index.js",
+  "typings": "dist/index.d.ts",
+  "typescript": {
+    "definition": "dist/index.d.ts"
+  },
+  "jest": {
+    "globals": {
+      "ts-jest": {
+        "enableTsDiagnostics": false
+      }
+    },
+    "transform": {
+      "^.+\\.tsx?$": "ts-jest"
+    },
+    "testRegex": "(/__tests__/.*|(\\.|/)(test|spec))\\.(jsx?|tsx?)$",
+    "moduleFileExtensions": [
+      "ts",
+      "tsx",
+      "js",
+      "jsx",
+      "json",
+      "node"
+    ]
+  }
+}
