+++ packages/templates/typescript-resolvers/src/after-schema.handlebars
@@ -1,0 +1,4 @@
+{{{ blockCommentIf 'Resolvers' types }}}
+{{#each types}}
+{{~> resolver }}
+{{/each}}
