+++ packages/templates/typescript-resolvers/src/index.ts
@@ -1,0 +1,3 @@
+import { config } from './config';
+
+export default config;
