+++ packages/templates/typescript-resolvers/src/polyfills.d.ts
@@ -1,0 +1,4 @@
+declare module '*.handlebars' {
+  const content: string;
+  export = content;
+}
